requirements/runtime.txt
PyYAML < 4.0.0
plumbum < 1.5.0 
click < 4.0.0
giturlparse.py < 0.1.0
future < 1.0.0
path.py < 5.0.0
python-magic < 0.5.0
requests < 2.5.0
semantic_version < 3.0.0
uritemplate < 1.0.0
twine < 1.4.0
wheel < 1.0.0
